Moved code related to ANSI escape codes to new file called ansi.py
