+++ examples/pirate.py
@@ -9,7 +9,6 @@ It demonstrates many features of cmd2.
 import argparse
 
 import cmd2
-from cmd2 import utils
 from cmd2.constants import MULTILINE_TERMINATOR
 
 
@@ -70,7 +69,7 @@ class Pirate(cmd2.Cmd):
 
     def do_sing(self, arg):
         """Sing a colorful song."""
-        self.poutput(utils.style(arg, fg=self.songcolor))
+        self.poutput(cmd2.ansi.style(arg, fg=self.songcolor))
 
     yo_parser = argparse.ArgumentParser()
     yo_parser.add_argument('--ho', type=int, default=2, help="How often to chant 'ho'")
