+++ tests/test_utils.py
@@ -10,6 +10,7 @@ import pytest
 
 from colorama import Fore, Back
 import cmd2.utils as cu
+import cmd2.ansi as ansi
 
 HELLO_WORLD = 'Hello, world!'
 
@@ -18,28 +19,28 @@ def test_strip_ansi():
     base_str = HELLO_WORLD
     ansi_str = Fore.GREEN + base_str + Fore.RESET
     assert base_str != ansi_str
-    assert base_str == cu.strip_ansi(ansi_str)
+    assert base_str == ansi.strip_ansi(ansi_str)
 
 def test_ansi_safe_wcswidth():
     base_str = HELLO_WORLD
     ansi_str = Fore.GREEN + base_str + Fore.RESET
-    assert cu.ansi_safe_wcswidth(ansi_str) != len(ansi_str)
+    assert ansi.ansi_safe_wcswidth(ansi_str) != len(ansi_str)
 
 def test_style():
     base_str = HELLO_WORLD
     ansi_str = Fore.BLUE + Back.GREEN + base_str + Fore.RESET + Back.RESET
-    assert cu.style(base_str, fg='blue', bg='green') == ansi_str
+    assert ansi.style(base_str, fg='blue', bg='green') == ansi_str
 
 def test_style_color_not_exist():
     base_str = HELLO_WORLD
     try:
-        cu.style(base_str, fg='hello', bg='green')
+        ansi.style(base_str, fg='hello', bg='green')
         assert False
     except ValueError:
         assert True
 
     try:
-        cu.style(base_str, fg='blue', bg='hello')
+        ansi.style(base_str, fg='blue', bg='hello')
         assert False
     except ValueError:
         assert True
